+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkNotSameWildcardsSetInspection.kt
@@ -1,18 +1,26 @@
 package com.jetbrains.snakecharm.inspections
 
 import com.intellij.codeInspection.LocalInspectionToolSession
-import com.intellij.codeInspection.ProblemHighlightType
+import com.intellij.codeInspection.ProblemHighlightType.WEAK_WARNING
 import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.Ref
+import com.intellij.psi.util.parentOfType
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.lang.psi.*
 
 class SmkNotSameWildcardsSetInspection : SnakemakeInspection() {
+    companion object {
+        val KEY = Key<HashMap<SmkRuleOrCheckpoint, Ref<List<WildcardDescriptor>>>>("SmkNotSameWildcardsSetInspection_Wildcards")
+    }
+
     override fun buildVisitor(
-            holder: ProblemsHolder,
-            isOnTheFly: Boolean,
-            session: LocalInspectionToolSession
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+        session: LocalInspectionToolSession
     ) = object : SnakemakeInspectionVisitor(holder, session) {
+        private val visitedRules = mutableSetOf<String>()
+
         override fun visitSmkRule(rule: SmkRule) {
             processRuleOrCheckpointLike(rule)
         }
@@ -21,43 +29,42 @@ class SmkNotSameWildcardsSetInspection : SnakemakeInspection() {
             processRuleOrCheckpointLike(checkPoint)
         }
 
+        override fun visitSmkUse(use: SmkUse) {
+            processRuleOrCheckpointLike(use)
+        }
+
         fun processRuleOrCheckpointLike(ruleOrCheckpoint: SmkRuleOrCheckpoint) {
-            var wildcardsRef: Ref<List<String>?>? = null
+            val cachedWildcardsByRule = session.putUserDataIfAbsent(KEY, hashMapOf())
+            val collector = AdvancedWildcardsCollector(
+                visitDefiningSections = true,
+                visitExpandingSections = false,
+                ruleLike = ruleOrCheckpoint,
+                cachedWildcardsByRule = cachedWildcardsByRule
+            )
 
-            ruleOrCheckpoint.getSections().forEach { section ->
-                if (section !is SmkRuleOrCheckpointArgsSection || !section.isWildcardsDefiningSection()) {
-                    return@forEach
-                }
-
-                if (wildcardsRef == null) {
-                    // Cannot do via types, we'd like to have wildcards only from
-                    // defining sections and ensure that defining sections could be parsed
-                    val collector = SmkWildcardsCollector(
-                            visitDefiningSections = true,
-                            visitExpandingSections = false
-                    )
-                    ruleOrCheckpoint.accept(collector)
-                    wildcardsRef = Ref.create(collector.getWildcardsNames())
-                }
-
-                val wildcards = wildcardsRef!!.get()
-                if (wildcards != null) {
-                    // show warnings only if wildcards defined
-                    processSection(section, wildcards)
-                }
+            val wildcards = collector.getDefinedWildcards().map { it.text }.toSet()
+            val visitedSections = mutableSetOf<String>()
+            handleSections(null, ruleOrCheckpoint, visitedSections, wildcards)
+            if (ruleOrCheckpoint is SmkUse) {
+                visitedRules.add(ruleOrCheckpoint.name ?: return)
+                ruleOrCheckpoint.getDefinedReferencesOfImportedRuleNames()
+                    ?.forEach { handleUseParent(ruleOrCheckpoint, it, visitedSections, wildcards) }
             }
         }
 
         private fun processSection(
-                section: SmkRuleOrCheckpointArgsSection,
-                wildcards: List<String>
+            section: SmkRuleOrCheckpointArgsSection,
+            wildcards: Set<String>,
+            originalUseParent: SmkUse?
         ) {
+            val sectionName = section.sectionKeyword ?: return
             val sectionArgs = section.argumentList?.arguments ?: return
             sectionArgs.forEach { arg ->
+                val errorTarget = originalUseParent?.nameIdentifier ?: arg
                 // collect wildcards in section arguments:
                 val collector = SmkWildcardsCollector(
-                        visitDefiningSections = false,
-                        visitExpandingSections = false
+                    visitDefiningSections = false,
+                    visitExpandingSections = false
                 )
                 arg.accept(collector)
                 val argWildcards = collector.getWildcardsNames()
@@ -66,26 +73,79 @@ class SmkNotSameWildcardsSetInspection : SnakemakeInspection() {
                     // if wildcards defined for args section
                     val missingWildcards = wildcards.filter { it !in argWildcards }
                     if (missingWildcards.isNotEmpty()) {
-                        registerProblem(
-                                arg,
-                                SnakemakeBundle.message(
-                                        "INSP.NAME.not.same.wildcards.set",
-                                        missingWildcards.sorted().joinToString() { "'$it'"}
-                                )
-                        )
+                        val message = if (originalUseParent != null) {
+                            SnakemakeBundle.message(
+                                "INSP.NAME.not.same.wildcards.set.in.parent.rule",
+                                missingWildcards.sorted().joinToString { "'$it'" },
+                                sectionName
+                            )
+                        } else {
+                            SnakemakeBundle.message(
+                                "INSP.NAME.not.same.wildcards.set",
+                                missingWildcards.sorted().joinToString { "'$it'" }
+                            )
+                        }
+                        registerProblem(errorTarget, message)
                     }
                 } else {
                     if (wildcards.isNotEmpty()) {
                         // no injections, cannot check
-                        registerProblem(
-                            arg,
+                        val message = if (originalUseParent != null) {
+                            SnakemakeBundle.message(
+                                "INSP.NAME.not.same.wildcards.set.cannot.check.in.parent.rule",
+                                sectionName
+                            )
+                        } else {
                             SnakemakeBundle.message(
                                 "INSP.NAME.not.same.wildcards.set.cannot.check"
-                            ),
-                            ProblemHighlightType.WEAK_WARNING
-                        )
+                            )
+                        }
+                        registerProblem(errorTarget, message, WEAK_WARNING)
                     }
                 }
+            }
+        }
+
+        private fun handleUseParent(
+            originalUseParent: SmkUse,
+            parentRef: SmkReferenceExpression,
+            visitedSections: MutableSet<String>,
+            wildcards: Set<String>
+        ) {
+            var resolveResult = parentRef.reference.resolve()
+            while (resolveResult is SmkReferenceExpression) {
+                val parentUseSection = resolveResult.parentOfType<SmkUse>() ?: return
+                handleSections(originalUseParent, parentUseSection, visitedSections, wildcards)
+                resolveResult = resolveResult.reference.resolve()
+            }
+            if (resolveResult !is SmkUse && resolveResult is SmkRuleOrCheckpoint) {
+                handleSections(originalUseParent, resolveResult, visitedSections, wildcards)
+                return
+            }
+            val useParent =
+                if (resolveResult is SmkUse) resolveResult else resolveResult?.parentOfType() ?: return
+            val useName = useParent.name ?: return
+            if(visitedRules.contains(useName)) {
+                return
+            }
+            visitedRules.add(useName)
+            handleSections(originalUseParent, useParent, visitedSections, wildcards)
+            useParent.getDefinedReferencesOfImportedRuleNames()
+                ?.forEach { handleUseParent(originalUseParent, it, visitedSections, wildcards) }
+        }
+
+        private fun handleSections(
+            originalUseParent: SmkUse?,
+            ruleLike: SmkRuleOrCheckpoint,
+            visitedSections: MutableSet<String>,
+            wildcards: Set<String>
+        ) {
+            ruleLike.getSections().forEach { section ->
+                if (section.sectionKeyword in visitedSections || section !is SmkRuleOrCheckpointArgsSection || !section.isWildcardsDefiningSection()) {
+                    return@forEach
+                }
+                visitedSections.add(section.sectionKeyword ?: return@forEach)
+                processSection(section, wildcards, originalUseParent)
             }
         }
     }