+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/smksl/SmkWildcardNotDefinedInspection.kt
@@ -7,85 +7,71 @@ import com.intellij.openapi.util.Key
 import com.intellij.openapi.util.Ref
 import com.jetbrains.snakecharm.SnakemakeBundle
 import com.jetbrains.snakecharm.inspections.SnakemakeInspection
-import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpoint
-import com.jetbrains.snakecharm.lang.psi.SmkRuleOrCheckpointArgsSection
-import com.jetbrains.snakecharm.lang.psi.SmkWildcardsCollector
-import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpressionImpl
+import com.jetbrains.snakecharm.lang.psi.*
+import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpression
 
 class SmkWildcardNotDefinedInspection : SnakemakeInspection() {
     companion object {
-        val KEY = Key<HashMap<SmkRuleOrCheckpoint, Ref<List<String>>>>("SmkWildcardNotDefinedInspection_Wildcards")
+        val KEY = Key<HashMap<SmkRuleOrCheckpoint, Ref<List<WildcardDescriptor>>>>("SmkWildcardNotDefinedInspection_Wildcards")
     }
 
     override fun buildVisitor(
-            holder: ProblemsHolder,
-            isOnTheFly: Boolean,
-            session: LocalInspectionToolSession
+        holder: ProblemsHolder,
+        isOnTheFly: Boolean,
+        session: LocalInspectionToolSession,
     ) = object : SmkSLInspectionVisitor(holder, session) {
 
-        override fun visitSmkSLReferenceExpression(expr: SmkSLReferenceExpressionImpl) {
+        override fun visitSmkSLReferenceExpression(expr: SmkSLReferenceExpression) {
             if (!expr.isWildcard()) {
                 return
             }
 
-            val ruleOrCheckpointSection = expr.containingRuleOrCheckpointSection() ?: return
-            val ruleOrCheckpoint = ruleOrCheckpointSection.getParentRuleOrCheckPoint()
+            val ruleLikeBlock = expr.containingRuleOrCheckpointSection()?.getParentRuleOrCheckPoint() ?: return
+            val cachedWildcardsByRule = session.putUserDataIfAbsent(KEY, hashMapOf())
+            val advancedCollector = AdvancedWildcardsCollector(
+                visitDefiningSections = true,
+                visitExpandingSections = false,
+                ruleLike = ruleLikeBlock,
+                cachedWildcardsByRule = cachedWildcardsByRule
+            )
+            val wildcards = advancedCollector.getDefinedWildcards().map { it.text }
+            val wildcardsCollectedInAllOverriddenRules = advancedCollector.wildcardsCollectedInAllOverriddenRules()
+            val inUseRuleBlock = ruleLikeBlock is SmkUse
 
-            val wildcardsByRule = session.putUserDataIfAbsent(KEY, hashMapOf())
-            if (ruleOrCheckpoint !in wildcardsByRule) {
-                val wildcardsDefiningSectionsAvailable = ruleOrCheckpoint.getSections()
-                        .asSequence()
-                        .filterIsInstance(SmkRuleOrCheckpointArgsSection::class.java)
-                        .filter { it.isWildcardsDefiningSection() }.firstOrNull() != null
-
-                val wildcards = when {
-                    // if no suitable sections let's think that no wildcards
-                    !wildcardsDefiningSectionsAvailable -> emptyList()
-                    else -> {
-                        // Cannot do via types, we'd like to have wildcards only from
-                        // defining sections and ensure that defining sections could be parsed
-                        val collector = SmkWildcardsCollector(
-                                visitDefiningSections = true,
-                                visitExpandingSections = false
-                        )
-                        ruleOrCheckpoint.accept(collector)
-                        collector.getWildcardsNames()
-//
-//                        val type = TypeEvalContext.codeAnalysis(expr.project, expr.containingFile)
-//                                .getType(ruleOrCheckpoint.wildcardsElement)
-//
-//                        var definedWildcards: List<WildcardDescriptor>? = null
-//                        if (type is SmkWildcardsType) {
-//                            definedWildcards = type.wildcardsDeclarations
-//                                    ?.filter { it.definingSectionRate != UNDEFINED_SECTION }
-//                                    ?.ifEmpty { null }
-//                        }
-//                        definedWildcards
-                    }
-                }
-                wildcardsByRule[ruleOrCheckpoint] =  Ref.create(wildcards)
+            // If an appropriate wildcard exists
+            if (expr.text in wildcards) {
+                return
             }
-            val wildcards = wildcardsByRule.getValue(ruleOrCheckpoint).get()
-            if (wildcards == null) {
+            // If no, firstly we need to check if there are failures in wildcard collecting
+            // If so, adds a weak warning
+            if (!wildcardsCollectedInAllOverriddenRules) {
                 // failed to parse wildcards defining sections
                 registerProblem(
-                        expr,
-                        SnakemakeBundle.message("INSP.NAME.wildcard.not.defined.cannot.check", expr.text),
-                        ProblemHighlightType.WEAK_WARNING
+                    expr,
+                    SnakemakeBundle.message("INSP.NAME.wildcard.not.defined.cannot.check", expr.text),
+                    ProblemHighlightType.WEAK_WARNING
                 )
             } else {
-                if (expr.text !in wildcards) {
-                    val definingSection = ruleOrCheckpoint.getWildcardDefiningSection()?.sectionKeyword
-                    val message = when (definingSection) {
-                        null -> SnakemakeBundle.message("INSP.NAME.wildcard.not.defined", expr.text)
-                        else -> SnakemakeBundle.message(
-                                "INSP.NAME.wildcard.not.defined.in.section",
-                                expr.text, definingSection
+                // Otherwise, adds an error
+                val definingSection = ruleLikeBlock.getWildcardDefiningSection()?.sectionKeyword
+                val message = when {
+                    inUseRuleBlock -> {
+                        SnakemakeBundle.message(
+                            "INSP.NAME.wildcard.not.defined.in.overridden.rule",
+                            expr.text
                         )
                     }
-
-                    registerProblem(expr, message)
+                    definingSection == null -> {
+                        SnakemakeBundle.message("INSP.NAME.wildcard.not.defined", expr.text)
+                    }
+                    else -> {
+                        SnakemakeBundle.message(
+                            "INSP.NAME.wildcard.not.defined.in.section",
+                            expr.text, definingSection
+                        )
+                    }
                 }
+                registerProblem(expr, message)
             }
         }
     }