+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkWildcardsCollector.kt
@@ -1,12 +1,17 @@
 package com.jetbrains.snakecharm.lang.psi
 
 import com.intellij.lang.injection.InjectedLanguageManager
+import com.intellij.openapi.util.Ref
+import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.parentOfType
 import com.jetbrains.python.psi.*
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.FUNCTIONS_BANNED_FOR_WILDCARDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_DEFINING_SECTIONS_KEYWORDS
+import com.jetbrains.snakecharm.lang.SnakemakeNames
 import com.jetbrains.snakecharm.stringLanguage.lang.callSimpleName
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLFile
+import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpression
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpressionImpl
 
 /**
@@ -18,15 +23,18 @@ import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLReferenceExpression
  *  @param visitDefiningSections Visit or not downstream sections which introduces new wildcards
  *  @param visitExpandingSections Visit or not downstream sections which allows wildcards usage w/o 'wildcards.' prefix
  *  @param visitAllSections If true visit all sections (including run) ignoring [visitDefiningSections] and [visitExpandingSections] flags
+ *  @param collectWildcardLikeReferences If true also collects [SmkSLReferenceExpression]s that are not wildcards
  */
 class SmkWildcardsCollector(
     private val visitDefiningSections: Boolean,
     private val visitExpandingSections: Boolean,
     private val visitAllSections: Boolean = false,
+    private val collectWildcardLikeReferences: Boolean = false
 ) : SmkElementVisitor, PyRecursiveElementVisitor() {
     private val wildcardsElements = mutableListOf<WildcardDescriptor>()
     private var atLeastOneInjectionVisited = false
     private var currentSectionIdx: Byte = -1
+    private var currentSectionName: String? = null
 
     /**
      * @return List of all wildcard element usages and its names or null if no string literals were found
@@ -44,6 +52,7 @@ class SmkWildcardsCollector(
 
     override fun visitSmkRunSection(st: SmkRunSection) {
         if (visitAllSections) {
+            currentSectionName = SnakemakeNames.SECTION_RUN
             super.visitSmkRunSection(st)
         }
     }
@@ -62,6 +71,7 @@ class SmkWildcardsCollector(
                 (visitDefiningSections && st.isWildcardsDefiningSection()) ||
                 (visitExpandingSections && st.isWildcardsExpandingSection())
             ) {
+                currentSectionName = st.sectionKeyword
                 super.visitSmkRuleOrCheckpointArgsSection(st)
                 return
             }
@@ -92,20 +102,199 @@ class SmkWildcardsCollector(
 
         val injections = PsiTreeUtil.getChildrenOfType(file, SmkSLReferenceExpressionImpl::class.java)
         injections?.forEach { st ->
-            wildcardsElements.add(WildcardDescriptor(
-                st, st.text,
-                if (currentSectionIdx == (-1).toByte()) WildcardDescriptor.UNDEFINED_SECTION_RATE else currentSectionIdx
-            ))
+            // We need to check whether is injection is wildcard because 'output' in:
+            // 'shell: "{output}"' is SmkSLReferenceExpressionImpl but not a wildcard
+            // We should try to resolve the reference because 'shell: "{wildcards.s}"'
+            // Is not a wildcard, but it refers to it
+            // On the same time we have to collect other references in some resolve/completion cases
+            // (see wildcards_resolve.feature 'Resolve first part of qualified like wildcard references into itself')
+            val potentialWildcardLike = if (collectWildcardLikeReferences || st.isWildcard()) st else {
+                (st.reference.resolve() as? SmkSLReferenceExpression)?.let { resolveResult ->
+                    if (resolveResult.isWildcard()) {
+                        resolveResult
+                    } else {
+                        null
+                    }
+                }
+            }
+            if (potentialWildcardLike != null) {
+                wildcardsElements.add(
+                    WildcardDescriptor(
+                        potentialWildcardLike, potentialWildcardLike.text,
+                        if (currentSectionIdx == (-1).toByte()) WildcardDescriptor.UNDEFINED_SECTION_RATE else currentSectionIdx,
+                        currentSectionName, st.isWildcard()
+                    )
+                )
+            }
         }
     }
 }
 
 data class WildcardDescriptor(
-    val psi: SmkSLReferenceExpressionImpl,
+    // TODO: consider using SmartPsiElementPointer instances instead of storing PSI elements directly or
+    //    at least check psi.isValid() before usage
+    val psi: SmkSLReferenceExpression,
     val text: String,
     val definingSectionRate: Byte,
+    val sectionName: String?,
+    val realWildcard: Boolean = false
 ) {
     companion object {
         const val UNDEFINED_SECTION_RATE: Byte = Byte.MAX_VALUE
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as WildcardDescriptor
+
+        if (text != other.text) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        return text.hashCode()
+    }
+}
+
+/**
+ * Manages collecting of wildcards from [ruleLike]. It is uses [cachedWildcardsByRule] to save and reads wildcards.
+ *
+ *  @param visitDefiningSections Visits or not downstream sections which introduces new wildcards
+ *  @param visitExpandingSections Visits or not downstream sections which allows wildcards usage w/o 'wildcards.' prefix
+ *  @param visitAllSections If true visits all sections (including run) ignoring [visitDefiningSections] and [visitExpandingSections] flags
+ *  @param getIntersection If true takes intersection of sets of inherited wildcards. Otherwise, takes union of them
+ *  @param collectWildcardLikeReferences If true also collects [SmkSLReferenceExpression]s that are not wildcards
+ */
+class AdvancedWildcardsCollector(
+    private val visitDefiningSections: Boolean,
+    private val visitExpandingSections: Boolean,
+    private val visitAllSections: Boolean = false,
+    private val getIntersection: Boolean = true,
+    private val ruleLike: SmkRuleOrCheckpoint,
+    private val cachedWildcardsByRule: HashMap<SmkRuleOrCheckpoint, Ref<List<WildcardDescriptor>>>?,
+    private val collectWildcardLikeReferences: Boolean = false
+) {
+    private val visitedWildcardDefinitionSections = mutableSetOf<String>()
+    private val wildcardsElementsDescriptors = mutableListOf<WildcardDescriptor>()
+    private var wildcardsCollectedInAllOverriddenRules = true
+    private val visitedRules = mutableSetOf<SmkRuleOrCheckpoint>()
+
+    fun wildcardsCollectedInAllOverriddenRules() = wildcardsCollectedInAllOverriddenRules
+
+    fun getDefinedWildcards(): List<WildcardDescriptor> {
+        // Collects wildcards, defined in current rule-like section
+        checkWildcardsAndUpdateThem(ruleLike, wildcardsElementsDescriptors, visitedWildcardDefinitionSections)
+        if (ruleLike is SmkUse) {
+            // Advanced collectIon of wildcards that were inherited
+            getWildcardsIntersection(ruleLike, wildcardsElementsDescriptors, visitedWildcardDefinitionSections)
+        }
+
+        return wildcardsElementsDescriptors
+    }
+
+    private fun collectWildcards(
+        ruleLike: PsiElement,
+        listOfDescriptors: MutableList<WildcardDescriptor>,
+        visitedSections: MutableSet<String>
+    ) {
+        // Trying to resolve the reference
+        var resolveResult = if (ruleLike is SmkReferenceExpression) ruleLike.reference.resolve() else ruleLike
+        while (resolveResult is SmkReferenceExpression) {
+            // Collects wildcards in each step of the resolving
+            val newUseRule = resolveResult.parentOfType<SmkUse>() ?: return
+            checkWildcardsAndUpdateThem(newUseRule, listOfDescriptors, visitedSections)
+            resolveResult = resolveResult.reference.resolve()
+        }
+        if (resolveResult in visitedRules) {
+            return
+        }
+        // Collects wildcards from current section if it just a rule or checkpoint
+        if (resolveResult !is SmkUse && resolveResult is SmkRuleOrCheckpoint) {
+            checkWildcardsAndUpdateThem(resolveResult, listOfDescriptors, visitedSections)
+            return
+        }
+        // Otherwise, it also produces wildcards collecting of inherited rules
+        val newUseRule =
+            (resolveResult as? SmkUse) ?: resolveResult?.parentOfType() ?: return
+        checkWildcardsAndUpdateThem(newUseRule, listOfDescriptors, visitedSections)
+        getWildcardsIntersection(newUseRule, listOfDescriptors, visitedSections)
+    }
+
+    private fun getWildcardsIntersection(
+        use: SmkUse,
+        listOfDescriptors: MutableList<WildcardDescriptor>,
+        visitedSections: MutableSet<String>
+    ) {
+        // Get lists of inherited wildcards
+        val wildcardsLists = (use.getDefinedReferencesOfImportedRuleNames()?.toList() ?: use.getImportedRules())
+            ?.map {
+                val listOfWildcards = mutableListOf<WildcardDescriptor>()
+                collectWildcards(it, listOfWildcards, visitedSections.toMutableSet())
+                listOfWildcards
+            }
+        var newDescriptors = wildcardsLists?.firstOrNull()?.toSet() ?: return
+        if (getIntersection) {
+            // Gets intersection in case of inspections
+            for (listOfWildcards in wildcardsLists) {
+                newDescriptors = newDescriptors intersect listOfWildcards.toSet()
+            }
+        } else {
+            // Gets union in case of resolve and completion
+            for (listOfWildcards in wildcardsLists) {
+                newDescriptors = newDescriptors union listOfWildcards
+            }
+        }
+        listOfDescriptors.addAll(newDescriptors)
+    }
+
+    private fun checkWildcardsAndUpdateThem(
+        ruleLike: SmkRuleOrCheckpoint,
+        listOfDescriptors: MutableList<WildcardDescriptor>,
+        visitedSections: MutableSet<String>
+    ) {
+        // Collects wildcards from current ruleLike section
+        visitedRules.add(ruleLike)
+        // Reads wildcards from the cache if possible
+        val descriptors = if (cachedWildcardsByRule != null && ruleLike in cachedWildcardsByRule) {
+            cachedWildcardsByRule.getValue(ruleLike).get()?.filter { it.psi.isValid }
+        } else {
+            // Otherwise, uses WildcardsCollector to gather wildcards
+            val wildcardsDefiningSectionsAvailable = ruleLike.getSections()
+                .asSequence()
+                .filterIsInstance(SmkRuleOrCheckpointArgsSection::class.java)
+                .filter { it.isWildcardsDefiningSection() }.firstOrNull() != null
+            val collector = SmkWildcardsCollector(
+                visitDefiningSections = visitDefiningSections,
+                visitExpandingSections = visitExpandingSections,
+                visitAllSections = visitAllSections,
+                collectWildcardLikeReferences = collectWildcardLikeReferences
+            )
+            ruleLike.accept(collector)
+            if (visitDefiningSections && !visitExpandingSections && !wildcardsDefiningSectionsAvailable) {
+                emptyList()
+            } else {
+                collector.getWildcards().also {
+                    if (cachedWildcardsByRule != null) {
+                        // Updates cache if possible
+                        cachedWildcardsByRule[ruleLike] = Ref.create(it)
+                    }
+                }
+            }
+        }
+        // Collects only those wildcards which were found in sections, that haven't been visited yet
+        descriptors?.filter { it.sectionName !in visitedSections }
+            ?.forEach { descriptor ->
+                listOfDescriptors.add(descriptor)
+            }
+        visitedSections.addAll(ruleLike.getSections().mapNotNull { it.sectionKeyword })
+
+        if (descriptors == null) {
+            // Highlights, that it was impossible to check some sections
+            wildcardsCollectedInAllOverriddenRules = false
+            return
+        }
     }
 }