+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/SmkUseImpl.kt
@@ -47,7 +47,7 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
             return namePattern.node
         }
         // There are no pattern or name node
-        val originalNames = getImportedRuleNames()
+        val originalNames = getDefinedReferencesOfImportedRuleNames()
         // Returns original names, we don't want to save one name because
         // index with this name probably already exists
         // so we save whole rules names if it is not just '*' wildcard
@@ -66,21 +66,12 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
         }
         val newName = getNameIdentifierPattern()
         val originalNames = mutableListOf<Pair<String, PsiElement>>()
-        getImportedRuleNames()?.forEach { reference ->
+        getDefinedReferencesOfImportedRuleNames()?.forEach { reference ->
             originalNames.add(reference.text to reference)
         }
         if (originalNames.isEmpty()) {
-            val moduleRef = (getModuleName() as? SmkReferenceExpression) ?: return emptyList()
-            val module = moduleRef.reference.resolve()
-            if (module == null || module !is SmkModule) {
-                return emptyList()
-            }
-            val file = module.getPsiFile()
-            if (file != null && file is SmkFile) {
-                originalNames.addAll(file.advancedCollectRules(visitedFiles).map { it.first to it.second })
-            } else {
-                return emptyList()
-            }
+            val pairs = getPairsOfImportedRulesAndNames(visitedFiles) ?: return emptyList()
+            originalNames.addAll(pairs.map { it.first to it.second })
         }
         return if (newName != null) {
             originalNames.map { newName.text.replace("*", it.first) to it.second }
@@ -94,13 +85,29 @@ class SmkUseImpl : SmkRuleLikeImpl<SmkUseStub, SmkUse, SmkRuleOrCheckpointArgsSe
      */
     private fun getNameIdentifierPattern() = findChildByType(SmkElementTypes.USE_NAME_IDENTIFIER) as? PsiElement
 
+    /**
+     * Collects all imported rules and their names
+     */
+    private fun getPairsOfImportedRulesAndNames(visitedFiles: MutableSet<PsiFile>): List<Pair<String, SmkRuleOrCheckpoint>>? {
+        val module = ((getModuleName() as? SmkReferenceExpression)?.reference?.resolve() as? SmkModule) ?: return null
+        val file = module.getPsiFile()
+        return if (file != null && file is SmkFile) {
+            file.advancedCollectRules(visitedFiles).map { it.first to it.second }
+        } else {
+            null
+        }
+    }
+
     override fun getModuleName() =
         (findChildByType(SmkTokenTypes.SMK_FROM_KEYWORD) as? PsiElement)?.nextSibling?.nextSibling
 
-    override fun getImportedRuleNames(): Array<SmkReferenceExpression>? = PsiTreeUtil.getChildrenOfType(
+    override fun getDefinedReferencesOfImportedRuleNames(): Array<SmkReferenceExpression>? = PsiTreeUtil.getChildrenOfType(
         findChildByType(USE_IMPORTED_RULES_NAMES),
         SmkReferenceExpression::class.java
     )
+
+    override fun getImportedRules(): List<SmkRuleOrCheckpoint>? =
+        getPairsOfImportedRulesAndNames(mutableSetOf())?.map { it.second }
 
     override fun nameIdentifierIsWildcard() = nameIdentifier?.let {
         it is SmkUseNameIdentifier && it.textContains('*')